disable flyway on development

+++ src/main/resources/db/postgresql/migration/V2__create_cms_engine.sql
@@ -1,6 +1,6 @@
 CREATE TABLE cms_articles (
   id         SERIAL PRIMARY KEY,
-  user_id    INT                         NOT NULL,
+  user_id    BIGINT                         NOT NULL,
   title      VARCHAR(255)                NOT NULL,
   summary    VARCHAR(800)                NOT NULL,
   body       TEXT                        NOT NULL,
@@ -24,16 +24,16 @@ CREATE UNIQUE INDEX idx_cms_tags_name
 
 CREATE TABLE cms_articles_tags (
   id         SERIAL PRIMARY KEY,
-  article_id INT NOT NULL,
-  tag_id     INT NOT NULL
+  article_id BIGINT NOT NULL,
+  tag_id     BIGINT NOT NULL
 );
 CREATE UNIQUE INDEX idx_cms_articles_tags
   ON cms_articles_tags (article_id, tag_id);
 
 CREATE TABLE cms_comments (
   id         SERIAL PRIMARY KEY,
-  article_id INT                         NOT NULL,
-  user_id    INT                         NOT NULL,
+  article_id BIGINT                         NOT NULL,
+  user_id    BIGINT                         NOT NULL,
   body       TEXT                        NOT NULL,
   type       VARCHAR(8)                  NOT NULL DEFAULT 'markdown',
   created_at TIMESTAMP WITHOUT TIME ZONE NOT NULL DEFAULT now(),
